# src/api/handlers/countries.go
package handlers

import (
	_ "CarSaleAd-Web-Api/api/dto"
	_ "CarSaleAd-Web-Api/api/helper"
	"CarSaleAd-Web-Api/config"
	"CarSaleAd-Web-Api/services"

	"github.com/gin-gonic/gin"
)

type CountryHandler struct {
	service *services.CountryService
}

func NewCountryHandler(cfg *config.Config) *CountryHandler {
	return &CountryHandler{service: services.NewCountryService(cfg)}
}

// CreateCountry godoc
// @Summary Create a Country
// @Description Create a Country
// @Tags Country
// @Accept json
// @Produces json
// @Param Request body dto.CreateUpdateCountryRequest true "Create a Country"
// @Success 201 {object} helper.BaseHttpResponse{result=dto.CountryResponse} "Country response"
// @Failure 400 {object} helper.BaseHttpResponse "Bad request"
// @Router /v1/countries/ [post]
// @Security AuthBearer
func (h *CountryHandler) Create(c *gin.Context) {
	Create(c, h.service.Create)
	// req := dto.CreateUpdateCountryRequest{}
	// err := c.ShouldBindJSON(&req)

	// if err != nil {
	// 	c.AbortWithStatusJSON(http.StatusBadRequest,
	// 		helper.GenerateBaseResponseWithValidationError(nil, false, 121, err))
	// 	return
	// }

	// res, err := h.service.Create(c, &req)

	// if err != nil {
	// 	c.AbortWithStatusJSON(helper.TranslateErrorToStatusCode(err),
	// 		helper.GenerateBaseResponseWithError(nil, false, 121, err))
	// 	return
	// }
	// c.JSON(http.StatusCreated, helper.GenerateBaseResponse(res, true, 0))

}

// UpdateCountry godoc
// @Summary Update a Country
// @Description Update a Country
// @Tags Country
// @Accept json
// @Produces json
// @Param id path int true "Id"
// @Param Request body dto.CreateUpdateCountryRequest true "Update a Country"
// @Success 200 {object} helper.BaseHttpResponse{result=dto.CountryResponse} "Country response"
// @Failure 400 {object} helper.BaseHttpResponse "Bad request"
// @Router /v1/countries/{id} [put]
// @Security AuthBearer
func (h *CountryHandler) Update(c *gin.Context) {
	Update(c, h.service.Update)
	// id, _ := strconv.Atoi(c.Params.ByName("id"))
	// req := dto.CreateUpdateCountryRequest{}
	// err := c.ShouldBindJSON(&req)

	// if err != nil {
	// 	c.AbortWithStatusJSON(http.StatusBadRequest,
	// 		helper.GenerateBaseResponseWithValidationError(nil, false, 121, err))
	// 	return
	// }

	// res, err := h.service.Update(c, id, &req)

	// if err != nil {
	// 	c.AbortWithStatusJSON(helper.TranslateErrorToStatusCode(err),
	// 		helper.GenerateBaseResponseWithError(nil, false, 121, err))
	// 	return
	// }
	// c.JSON(http.StatusOK, helper.GenerateBaseResponse(res, true, 0))

}

// DeleteCountry godoc
// @Summary Delete a Country
// @Description Delete a Country
// @Tags Country
// @Accept json
// @Produces json
// @Param id path int true "Id"
// @Success 200 {object} helper.BaseHttpResponse "Response"
// @Failure 400 {object} helper.BaseHttpResponse "Bad request"
// @Router /v1/countries/{id} [delete]
// @Security AuthBearer
func (h *CountryHandler) Delete(c *gin.Context) {
	Delete(c, h.service.Delete)
	// id, _ := strconv.Atoi(c.Params.ByName("id"))
	// if id == 0 {
	// 	c.AbortWithStatusJSON(http.StatusNotFound,
	// 		helper.GenerateBaseResponse(nil, false, 121))
	// 	return

	// }

	// err := h.service.Delete(c, id)
	// if err != nil {
	// 	c.AbortWithStatusJSON(helper.TranslateErrorToStatusCode(err),
	// 		helper.GenerateBaseResponseWithError(nil, false, 121, err))
	// 	return
	// }

	// c.JSON(http.StatusOK, helper.GenerateBaseResponse(nil, true, 0))

}

// GetCountry godoc
// @Summary Get a Country
// @Description Get a Country
// @Tags Country
// @Accept json
// @Produces json
// @Param id path int true "Id"
// @Success 200 {object} helper.BaseHttpResponse{result=dto.CountryResponse} "Country Response"
// @Failure 400 {object} helper.BaseHttpResponse "Bad request"
// @Router /v1/countries/{id} [get]
// @Security AuthBearer
func (h *CountryHandler) GetById(c *gin.Context) {
	GetById(c, h.service.GetById)
	// id, _ := strconv.Atoi(c.Params.ByName("id"))
	// if id == 0 {
	// 	c.AbortWithStatusJSON(http.StatusNotFound,
	// 		helper.GenerateBaseResponse(nil, false, 121))
	// 	return
	// }

	// res, err := h.service.GetById(c, id)

	// if err != nil {
	// 	c.AbortWithStatusJSON(helper.TranslateErrorToStatusCode(err),
	// 		helper.GenerateBaseResponseWithError(nil, false, 121, err))
	// 	return
	// }

	// c.JSON(http.StatusOK, helper.GenerateBaseResponse(res, true, 0))

}

// GetCountries godoc
// @Summary Get Countries
// @Description Get Countries
// @Tags Country
// @Accept json
// @Produces json
// @Param Request body dto.PaginationInputWithFilter true "Request"
// @Success 200 {object} helper.BaseHttpResponse{result=dto.PagedList[dto.CountryResponse]} "Country Response"
// @Failure 400 {object} helper.BaseHttpResponse "Bad request"
// @Router /v1/countries/get-by-filter [post]
// @Security AuthBearer
func (h *CountryHandler) GetByFilter(c *gin.Context) {
	GetByFilter(c, h.service.GetByFilter)
	// req := dto.PaginationInputWithFilter{}
	// err := c.ShouldBindJSON(&req)

	// if err != nil {
	// 	c.AbortWithStatusJSON(http.StatusBadRequest,
	// 		helper.GenerateBaseResponseWithValidationError(nil, false, 121, err))
	// 	return
	// }

	// res, err := h.service.GetByFilter(c, &req)

	// if err != nil {
	// 	c.AbortWithStatusJSON(helper.TranslateErrorToStatusCode(err),
	// 		helper.GenerateBaseResponseWithError(nil, false, 121, err))
	// 	return
	// }

	// c.JSON(http.StatusOK, helper.GenerateBaseResponse(res, true, 0))

}