# src/docs/swagger.yaml
definitions:
  CarSaleAd-Web-Api_api_dto.CityResponse:
    properties:
      country:
        $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CountryResponse'
      id:
        type: integer
      name:
        type: string
    type: object
  CarSaleAd-Web-Api_api_dto.CountryResponse:
    properties:
      cities:
        items:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CityResponse'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
  CarSaleAd-Web-Api_api_dto.CreateUpdateCityRequest:
    properties:
      countryId:
        type: integer
      name:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - name
    type: object
  CarSaleAd-Web-Api_api_dto.CreateUpdateCountryRequest:
    properties:
      name:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  CarSaleAd-Web-Api_api_dto.Filter:
    properties:
      filterType:
        type: string
      from:
        type: string
      to:
        type: string
      type:
        type: string
    type: object
...
  CarSaleAd-Web-Api_api_dto.PagedList-CarSaleAd-Web-Api_api_dto_CityResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CityResponse'
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  CarSaleAd-Web-Api_api_dto.PagedList-CarSaleAd-Web-Api_api_dto_CountryResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CountryResponse'
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  CarSaleAd-Web-Api_api_dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.Filter'
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.Sort'
        type: array
    type: object
...
  CarSaleAd-Web-Api_api_dto.Sort:
    properties:
      col_id:
        type: string
      sort:
        type: string
    type: object
...
  CarSaleAd-Web-Api_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_validations.ValidationError'
        type: array
    type: object
  CarSaleAd-Web-Api_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  CarSaleAd-Web-Api_api_validations.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /v1/cities/:
    post:
      consumes:
      - application/json
      description: Create a City
      parameters:
      - description: Create a City
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CreateUpdateCityRequest'
      responses:
        "201":
          description: City response
          schema:
            allOf:
            - $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a City
      tags:
      - City
  /v1/cities/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a City
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Response
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a City
      tags:
      - City
    get:
      consumes:
      - application/json
      description: Get a City
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: City Response
          schema:
            allOf:
            - $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a City
      tags:
      - City
    put:
      consumes:
      - application/json
      description: Update a City
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a City
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CreateUpdateCityRequest'
      responses:
        "200":
          description: City response
          schema:
            allOf:
            - $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a City
      tags:
      - City
  /v1/cities/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get Cities
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.PaginationInputWithFilter'
      responses:
        "200":
          description: City Response
          schema:
            allOf:
            - $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.PagedList-CarSaleAd-Web-Api_api_dto_CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Cities
      tags:
      - City
  /v1/countries/:
    post:
      consumes:
      - application/json
      description: Create a Country
      parameters:
      - description: Create a Country
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CreateUpdateCountryRequest'
      responses:
        "201":
          description: Country response
          schema:
            allOf:
            - $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CountryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a Country
      tags:
      - Country
  /v1/countries/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Country
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Response
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a Country
      tags:
      - Country
    get:
      consumes:
      - application/json
      description: Get a Country
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Country Response
          schema:
            allOf:
            - $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CountryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a Country
      tags:
      - Country
    put:
      consumes:
      - application/json
      description: Update a Country
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a Country
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CreateUpdateCountryRequest'
      responses:
        "200":
          description: Country response
          schema:
            allOf:
            - $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.CountryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a Country
      tags:
      - Country
  /v1/countries/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get Countries
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.PaginationInputWithFilter'
      responses:
        "200":
          description: Country Response
          schema:
            allOf:
            - $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.PagedList-CarSaleAd-Web-Api_api_dto_CountryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Countries
      tags:
      - Country
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"